+++ src/bewegung/core/vector/array2d.py
@@ -8,7 +8,7 @@ https://github.com/pleiszenburg/bewegung
 
     src/bewegung/core/vector/array2d.py: 2D Vector Array
 
-    Copyright (C) 2020 Sebastian M. Ernst <ernst@pleiszenburg.de>
+    Copyright (C) 2020-2021 Sebastian M. Ernst <ernst@pleiszenburg.de>
 
 <LICENSE_BLOCK>
 The contents of this file are subject to the GNU Lesser General Public License
@@ -112,29 +112,39 @@ class VectorArray2D(VectorArray2DABC):
 
         return np.allclose(self.x, other.x) and np.allclose(self.y, other.y)
 
-    def __add__(self, other: VectorArray2DABC) -> VectorArray2DABC:
+    def __add__(self, other: Union[VectorArray2DABC, Vector2D]) -> VectorArray2DABC:
         """
-        Add operation between vector arrays
+        Add operation between vector arrays or a vector array and a vector
 
         Args:
             other : Another vector array of equal length
         """
 
-        assert len(self) == len(other)
-        assert self.dtype == other.dtype
+        if isinstance(other, VectorArray2DABC):
+            assert len(self) == len(other)
+            assert self.dtype == other.dtype
         return VectorArray2D(self.x + other.x, self.y + other.y)
 
-    def __sub__(self, other: VectorArray2DABC) -> VectorArray2DABC:
+    def __radd__(self, *args, **kwargs):
+
+        return self.__add__(*args, **kwargs)
+
+    def __sub__(self, other: Union[VectorArray2DABC, Vector2D]) -> VectorArray2DABC:
         """
-        Substract operator between vector arrays
+        Substract operator between vector arrays or a vector array and a vector
 
         Args:
             other : Another vector array of equal length
         """
 
-        assert len(self) == len(other)
-        assert self.dtype == other.dtype
+        if isinstance(other, VectorArray2DABC):
+            assert len(self) == len(other)
+            assert self.dtype == other.dtype
         return VectorArray2D(self.x - other.x, self.y - other.y)
+
+    def __rsub__(self, *args, **kwargs):
+
+        return self.__sub__(*args, **kwargs)
 
     def __mul__(self, other: Number) -> VectorArray2DABC:
         """
@@ -197,7 +207,7 @@ class VectorArray2D(VectorArray2DABC):
         Exports vector array as a tuple of polar vector components in ``numpy.ndarry`` objects
         """
 
-        return self.mag, np.arctan2(self._y, self._x)
+        return self.mag, self.angle
 
     def as_tuple(self) -> Tuple[np.ndarray, np.ndarray]:
         """
@@ -251,6 +261,14 @@ class VectorArray2D(VectorArray2DABC):
         """
 
         return np.sqrt(self._x ** 2 + self._y ** 2)
+
+    @property
+    def angle(self) -> np.ndarray:
+        """
+        The vectors' angles in radians, computed on demand
+        """
+
+        return np.arctan2(self._y, self._x)
 
     @property
     def x(self) -> np.ndarray: